+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Added QueryStringBuilder to help in building query strings.
+- Added FluentUriBuilder to help in building Uris.
 
 ### Changed